+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- Add the Main contact task to the Conversion project task list
+
 ### Changed
 
 - Remove the Funding agreement letters contact task from the Conversion project